fix float-int change

+++ q01_10/002.py
@@ -4,7 +4,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori.jpg")
-    ans = lib0110.BGR2GRAY(img).astype(np.uint8)
+    ans = lib0110.BGR2GRAY(img)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
