+++ q01_10/010.py
@@ -4,7 +4,7 @@ import lib0110
 
 if __name__ == '__main__':
     img = cv2.imread("Gasyori100knock/Question_01_10/imori_noise.jpg")
-    ans = lib0110.Median_filter(img).astype(np.uint8)
+    ans = lib0110.Median_filter(img)
 
     cv2.imshow("imori", ans)
     cv2.waitKey(0)
